+++ src/index.js
@@ -6,6 +6,17 @@ class pdfmake extends pdfmakeBase {
 		super();
 	}
 
+	/**
+	 * @param {(path: string) => boolean} callback
+	 */
+	setLocalAccessPolicy(callback) {
+		if (callback !== undefined && typeof callback !== 'function') {
+			throw new Error("Parameter 'callback' has an invalid type. Function or undefined expected.");
+		}
+
+		this.localAccessPolicy = callback;
+	}
+
 	_transformToDocument(doc) {
 		return new OutputDocumentServer(doc);
 	}